Rename Chatbot component to .js and update imports
